Post-integration cleanup: Fix for variable freezing in nested ForEach loops
- change m_FreezeCount to uint8_t to avoid increasing size of
BFFVariable
- add regression test for nested ForEach loops not being frozen
correctly

+++ Code/Tools/FBuild/FBuildCore/BFF/BFFVariable.h
@@ -92,7 +92,7 @@ private:
     AString m_Name;
     VarType m_Type;
 
-    mutable uint32_t m_FreezeCount;
+    mutable uint8_t     m_FreezeCount;
 
     //
     bool                m_BoolValue;
